# docs/v3/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "contact" : {
      "email" : "support@swagger.io",
      "name" : "API Support",
      "url" : "http://www.swagger.io/support"
    },
    "description" : "This will take in user requests for equipments they want",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "title" : "Equipment Picker Calculator API",
    "version" : "0.1"
  },
  "servers" : [ {
    "url" : "//localhost:8080/api"
  } ],
  "paths" : {
    "/:userId/accounts" : {
      "get" : {
        "description" : "Creates an account attached to a user",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/controller.GetAccountsForUserResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Creates an account attached to a user",
        "tags" : [ "identity" ]
      }
    },
    "/:userId/create-account" : {
      "post" : {
        "description" : "Creates an account attached to a user",
        "responses" : {
          "202" : {
            "content" : { },
            "description" : "Accepted"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Creates an account attached to a user",
        "tags" : [ "identity" ]
      }
    },
    "/create-user" : {
      "post" : {
        "description" : "Creates a user",
        "responses" : {
          "202" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "Accepted"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Creates a user",
        "tags" : [ "identity" ]
      }
    }
  },
  "components" : {
    "schemas" : {
      "controller.AccountResponse" : {
        "properties" : {
          "account_type" : {
            "type" : "string"
          },
          "currency" : {
            "type" : "string"
          },
          "date_created" : {
            "example" : "2024-02-29",
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "controller.GetAccountsForUserResponse" : {
        "properties" : {
          "accounts" : {
            "items" : {
              "$ref" : "#/components/schemas/controller.AccountResponse"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      }
    }
  },
  "x-original-swagger-version" : "2.0"
}